renamed django-mongodb -> django-mongodb-backend
