+++ .github/workflows/mongodb_settings.py
@@ -1,14 +1,14 @@
 DATABASES = {
     "default": {
-        "ENGINE": "django_mongodb",
+        "ENGINE": "django_mongodb_backend",
         "NAME": "djangotests",
     },
     "other": {
-        "ENGINE": "django_mongodb",
+        "ENGINE": "django_mongodb_backend",
         "NAME": "djangotests-other",
     },
 }
-DEFAULT_AUTO_FIELD = "django_mongodb.fields.ObjectIdAutoField"
+DEFAULT_AUTO_FIELD = "django_mongodb_backend.fields.ObjectIdAutoField"
 PASSWORD_HASHERS = ("django.contrib.auth.hashers.MD5PasswordHasher",)
 SECRET_KEY = "django_tests_secret_key"
 USE_TZ = False
