+++ docs/source/forms.rst
@@ -1,9 +1,9 @@
 Forms API reference
 ===================
 
-.. module:: django_mongodb.forms
+.. module:: django_mongodb_backend.forms
 
-Some MongoDB-specific fields are available in ``django_mongodb.forms``.
+Some MongoDB-specific fields are available in ``django_mongodb_backend.forms``.
 
 ``ObjectIdField``
 -----------------
