+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["hatchling", "hatch-requirements-txt>=0.4.1"]
 build-backend = "hatchling.build"
 
 [project]
-name = "django-mongodb"
+name = "django-mongodb-backend"
 dynamic = ["version", "dependencies"]
 description = "MongoDB backend for Django"
 readme = "README.md"
@@ -37,12 +37,12 @@ docs = [ "sphinx>=7"]
 
 [project.urls]
 Homepage = "https://www.mongodb.org"
-Documentation = "https://django-mongodb.readthedocs.io"
+Documentation = "https://django-mongodb-backend.readthedocs.io"
 Source = "https://github.com/mongodb-labs/django-mongodb"
 Tracker = "https://github.com/mongodb-labs/django-mongodb/issues"
 
 [tool.hatch.version]
-path = "django_mongodb/__init__.py"
+path = "django_mongodb_backend/__init__.py"
 validate-bump = false
 
 [tool.hatch.metadata.hooks.requirements_txt]
